+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.5.0] - 26 June 2019.
+
+* Refactor the code to use RPC instead of REST for Firestore.
+* Implement listening to document real-time updates.
+
 ## [0.1.0] - 26 June 2019.
 
 * Fix issues identified in the dart pub analysis.